small tweaks for training/tokenization

+++ experiments/simple_train_config.py
@@ -1,6 +1,4 @@
 from dataclasses import dataclass
-from typing import Literal, TypeAlias
-
 
 
 @dataclass(frozen=True)
@@ -11,6 +9,6 @@ class SimpleTrainConfig:
     learning_rate: float
     weight_decay: float | None = None
     min_lr_ratio: float | None = None
-    warmup: int | None = None
+    warmup: float | None = None
     cooldown: float | None = None
     z_loss_weight: float | None = None
